Island: Move setup_island_event_handlers to monkey_island.cc.setup
