+++ monkey/monkey_island/cc/__init__.py
@@ -1,2 +1,1 @@
 from .version import Version
-from .setup_island_event_handlers import setup_island_event_handlers
