+++ monkey/monkey_island/cc/setup/island_event_handlers.py
@@ -1,6 +1,6 @@
 from common import DIContainer
 from monkey_island.cc.event_queue import IslandEventTopic, PyPubSubIslandEventQueue
-from monkey_island.cc.services.reset_agent_configuration import reset_agent_configuration
+from monkey_island.cc.island_event_handlers import reset_agent_configuration
 
 
 def setup_island_event_handlers(container: DIContainer):
